🏷️(backoffice) remove types duplication
- Updated multiple hooks to import ResourcesQuery from
"@/hooks/useResources/types" instead of directly from "@/hooks/useResources".
- Cleaned up imports in various components and repositories to enhance code
organization and maintainability.
- Introduced a new `PaginatedResponse` interface in "@/types/api" for better
type management across repositories.
