+++ src/frontend/admin/src/components/presentational/table/usePaginatedTableResource.ts
@@ -3,7 +3,8 @@ import { GridPaginationModel } from "@mui/x-data-grid";
 import { useDebouncedCallback } from "use-debounce";
 import { keepPreviousData } from "@tanstack/react-query";
 import { useRouter } from "next/router";
-import { Resource, ResourcesQuery, useResources } from "@/hooks/useResources";
+import { Resource, useResources } from "@/hooks/useResources";
+import { ResourcesQuery } from "@/hooks/useResources/types";
 import { Maybe } from "@/types/utils";
 import { DEFAULT_PAGE_SIZE, DEFAULT_SEARCH_DEBOUNCE } from "@/utils/constants";
 import { deleteUnusedFilters, getAvailableFilters } from "@/utils/filters";
