+++ src/frontend/admin/src/hooks/useCourseProductRelation/useCourseProductRelation.ts
@@ -2,11 +2,11 @@ import { defineMessages, useIntl } from "react-intl";
 import { useMutation } from "@tanstack/react-query";
 import {
   QueryOptions,
-  ResourcesQuery,
   useResource,
   useResourcesCustom,
   UseResourcesProps,
 } from "@/hooks/useResources";
+import { ResourcesQuery } from "@/hooks/useResources/types";
 import { CourseProductRelationRepository } from "@/services/repositories/course-product-relation/CourseProductRelationRepository";
 import { CourseProductRelation } from "@/services/api/models/Relations";
 import { DTOOrderGroup } from "@/services/api/models/OrderGroup";
