+++ src/frontend/admin/src/services/repositories/AbstractRepository.ts
@@ -1,12 +1,6 @@
 import { ResourcesQuery } from "@/hooks/useResources";
-import { Maybe, Nullable } from "@/types/utils";
-
-export type PaginatedResponse<T> = {
-  count: number;
-  next: Nullable<number>;
-  previous: Nullable<number>;
-  results: T[];
-};
+import { PaginatedResponse } from "@/types/api";
+import { Maybe } from "@/types/utils";
 
 export interface AbstractRepository<
   T,
@@ -14,12 +8,8 @@ export interface AbstractRepository<
   DTOData,
 > {
   getAll: (filters?: Maybe<Filters>) => Promise<PaginatedResponse<T>>;
-
   get: (id: string, filters?: Maybe<Filters>) => Promise<T>;
-
   create: (payload: DTOData) => Promise<T>;
-
   update: (id: string, payload: DTOData) => Promise<T>;
-
   delete: (id: string) => Promise<void>;
 }
