+++ src/frontend/admin/src/services/repositories/orders/OrderRepository.ts
@@ -3,7 +3,7 @@ import { Order, OrderQuery } from "@/services/api/models/Order";
 import { Maybe } from "@/types/utils";
 import { ResourcesQuery } from "@/hooks/useResources/types";
 import { checkStatus, fetchApi } from "@/services/http/HttpService";
-import { PaginatedResponse } from "@/services/repositories/AbstractRepository";
+import { PaginatedResponse } from "@/types/api";
 
 export const orderRoutes = {
   get: (id: string, params: string = "") => `/orders/${id}/${params}`,
