+++ src/frontend/admin/src/types/api.ts
@@ -1,0 +1,8 @@
+import { Nullable } from "./utils";
+
+export interface PaginatedResponse<T> {
+  count: number;
+  next: Nullable<string>;
+  previous: Nullable<string>;
+  results: Array<T>;
+}
